+++ integrations/elasticsearch/src/haystack_integrations/components/retrievers/elasticsearch/__init__.py
@@ -4,6 +4,7 @@
 from .bm25_retriever import ElasticsearchBM25Retriever
 from .elasticsearch_hybrid_retriever import ElasticsearchHybridRetriever
 from .embedding_retriever import ElasticsearchEmbeddingRetriever
+from .inference_hybrid_retriever import ElasticsearchInferenceHybridRetriever
 from .inference_sparse_retriever import ElasticsearchInferenceSparseRetriever
 from .sparse_embedding_retriever import ElasticsearchSparseEmbeddingRetriever
 from .sql_retriever import ElasticsearchSQLRetriever
@@ -12,6 +13,7 @@ __all__ = [
     "ElasticsearchBM25Retriever",
     "ElasticsearchEmbeddingRetriever",
     "ElasticsearchHybridRetriever",
+    "ElasticsearchInferenceHybridRetriever",
     "ElasticsearchInferenceSparseRetriever",
     "ElasticsearchSQLRetriever",
     "ElasticsearchSparseEmbeddingRetriever",